broke campaigncreator into separate files
