+++ client/src/components/static_data_file_configs/ResourceListConfig.tsx
@@ -1,0 +1,5 @@
+import rawResourceListData from '../../static_data/resource_list.json';
+
+// --- Resource List Json ---
+export type TypeResourceListData = { [K in keyof typeof rawResourceListData]: string[] };
+export const resourceListData: TypeResourceListData = rawResourceListData;
